refactor(api): move user handlers to new package
